requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-on-heroku==1.1.2
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
protobuf==4.21.6
psycopg2-binary==2.9.3
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
validate-email-address==1
whitenoise==6.2.0
